% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{get_filterable_attributes}
\alias{get_filterable_attributes}
\title{Extract filterable attributes from a BIDS list}
\usage{
get_filterable_attributes(list)
}
\arguments{
\item{list}{The list of PET data or blood data}
}
\value{
A data.frame containing the filterable attributes
}
\description{
This function extracts the attributes which could be used for filtering. By
this, it extracts all parts of the list which are either character or numeric
which also have a length of 1. This allows us to filter our data for a
specific tracer, or injection type.
}
