% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_aif_preds}
\alias{plot_aif_preds}
\title{Plot arterial input function predictions}
\usage{
plot_aif_preds(data, preds, title)
}
\arguments{
\item{data}{AIF data}

\item{preds}{predictions dataframe}

\item{title}{plot title}
}
\value{
plots
}
\description{
Plot arterial input function predictions
}
\examples{
\dontrun{
plot_aif_preds(blooddata, preds, titletext)
}
}
