% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_blooddata_extra}
\alias{plot_blooddata_extra}
\title{Plot blooddata with an extra log time plot}
\usage{
plot_blooddata_extra(blooddata, titletext)
}
\arguments{
\item{blooddata}{blooddata object}

\item{titletext}{text of the title}
}
\value{
plots
}
\description{
Plot blooddata with an extra log time plot
}
\examples{
\dontrun{
plot_blooddata_extra(blooddata, titletext)
}
}
