% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bpr_preds}
\alias{plot_bpr_preds}
\title{Plot blood-to-plasma ratio predictions}
\usage{
plot_bpr_preds(data, preds, title)
}
\arguments{
\item{data}{BPR data}

\item{preds}{predictions dataframe}

\item{title}{plot title}
}
\value{
plots
}
\description{
Plot blood-to-plasma ratio predictions
}
\examples{
\dontrun{
plot_bpr_preds(blooddata, preds, titletext)
}
}
