% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pf_preds}
\alias{plot_pf_preds}
\title{Plot parent fraction predictions}
\usage{
plot_pf_preds(data, preds, title)
}
\arguments{
\item{data}{parent fraction data}

\item{preds}{predictions dataframe}

\item{title}{plot title}
}
\value{
plots
}
\description{
Plot parent fraction predictions
}
\examples{
\dontrun{
plot_pf_preds(blooddata, preds, titletext)
}
}
