% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_wb_preds}
\alias{plot_wb_preds}
\title{Plot whole-blood predictions}
\usage{
plot_wb_preds(data, preds, title)
}
\arguments{
\item{data}{whole blood data}

\item{preds}{predictions dataframe}

\item{title}{plot title}
}
\value{
plots
}
\description{
Plot whole-blood predictions
}
\examples{
\dontrun{
plot_aif_preds(blooddata, preds, titletext)
}
}
