% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bloodstream.R
\name{bloodstream}
\alias{bloodstream}
\title{Run the bloodstream pipeline}
\usage{
bloodstream(studypath, configpath = NULL)
}
\arguments{
\item{studypath}{The path to the BIDS study folder.}

\item{configpath}{The path to the config file.}
}
\value{
The derivatives files, and report will be produced in the study folder.
}
\description{
Run the bloodstream pipeline
}
\examples{
\dontrun{
bloodstream(studypath, configpath)
}
}
