% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bloodstream.R
\name{bloodstream}
\alias{bloodstream}
\title{Run the bloodstream pipeline}
\usage{
bloodstream(
  bids_dir,
  configpath = NULL,
  derivatives_dir = NULL,
  analysis_foldername = "Primary_Analysis",
  template_path = NULL
)
}
\arguments{
\item{bids_dir}{The path to the BIDS study folder.}

\item{configpath}{The path to the config file.}

\item{derivatives_dir}{The path to the derivatives directory. If NULL, uses bids_dir/derivatives.}

\item{analysis_foldername}{The name for the analysis subfolder (default: "Primary_Analysis").}

\item{template_path}{The path to the R Markdown template. If NULL, uses package default.}
}
\value{
The derivatives files, and report will be produced in the study folder.
}
\description{
Run the bloodstream pipeline
}
\examples{
\dontrun{
bloodstream(bids_dir, configpath)
bloodstream(bids_dir, configpath, analysis_foldername = "my_analysis")
}
}
