% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{generate_methods_boilerplate}
\alias{generate_methods_boilerplate}
\title{Generate methods boilerplate text for manuscript reporting}
\usage{
generate_methods_boilerplate(config, package_version = NULL)
}
\arguments{
\item{config}{The bloodstream configuration list}

\item{package_version}{The version of bloodstream used}
}
\value{
A character string containing the formatted methods text
}
\description{
Generate methods boilerplate text for manuscript reporting
}
